+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# react-components NEXT
+
+- [Icon] Add `star-empty` and `star-filled` icons  (by [@Sigler](https://github.com/Sigler) in [#303](https://github.com/puppetlabs/design-system/pull/303))
+
 # react-components 5.22.4 (2020-09-16)
 
 - [Icon] Add `azure`, `bitbucket`, `github`, and `gitlab` icons  (by [@eputnam](https://github.com/eputnam) in [#301](https://github.com/puppetlabs/design-system/pull/301))